Break status out into a module
